+++ NEWS.md
@@ -1,8 +1,9 @@
-# RPresto 1.3.2.9000
+# RPresto 1.3.3
 
 - Fix tests for compatibility with `dbplyr` 1.4.0.
 - Send headers with all http requests (#103).
 - Add support for `as.<data_type>` style casting for dplyr (#97).
+- Add Code of Conduct.
 
 # RPresto 1.3.2